+++ samples/CG.Business.QuickStart/Program.cs
@@ -1,5 +1,5 @@
-﻿using Microsoft.Extensions.Configuration;
-using Microsoft.Extensions.DependencyInjection;
+﻿using Microsoft.AspNetCore.Hosting;
+using Microsoft.Extensions.Hosting;
 using System;
 
 namespace CG.Business.QuickStart
@@ -8,19 +8,13 @@ namespace CG.Business.QuickStart
     {
         static void Main(string[] args)
         {
-            // Build a typical configuration.
-            var builder = new ConfigurationBuilder();
-            builder.AddJsonFile("appSettings.json");
-            var configuration = builder.Build();
-
-            // Build an empty DI container.
-            var services = new ServiceCollection();
-
-            // This is how to add repositories, using the configuration
-            //   to control what repository types get registered, and what
-            //   options are passed to those repositories, at runtime.
-            services.AddRepositories(configuration);
-
-        }            
+            Host.CreateDefaultBuilder(args)
+                .ConfigureWebHostDefaults(webBuilder =>
+                {
+                    webBuilder.UseStartup<Startup>(); // < -- call our startup class ...
+                })
+                .Build()
+                .Start();
+        } 
     }
 }