+++ samples/CG.Business.QuickStart/ServiceCollectionExtensions.cs
@@ -9,8 +9,12 @@ namespace Microsoft.Extensions.DependencyInjection
             IConfiguration configuration
             )
         {
-            // TODO : register concrete repository type here.
-                        
+            // This method is an example of how to hook into the startup
+            //   pipeline and run logic to register you specific repository
+            //   types, and, of course, any types those repositories also 
+            //   rely on. For instance, here is a gret place to register 
+            //   you data-context type(s), if you're using EFCore.
+
             return serviceCollection;
         }
     }